WIP: API Code Cleanup
Clean up API code to be reading from source folder instead of in the router
