+++ src/codecarto/containers/processor/api/routers/parser_router.py
@@ -1,7 +1,7 @@
-import re 
 import httpx
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter
 
+from src.util.exceptions import GithubError
 from api.util import generate_return, proc_exception, proc_error
 
 # DEBUG
@@ -20,10 +20,24 @@ async def parse():
 
 @ParserRoute.get("/handle_github_url")
 async def handle_github_url(github_url: str) -> dict:
+    """Handles a GitHub URL and returns a dictionary of directories and files
+
+    Parameters:
+        github_url {str} -- GitHub URL to handle
+
+    Returns:
+        dict -- Dictionary of directories and files
+    """
     import time
+    from src.parser.import_source_url import (
+        ImportSourceUrlError,
+        read_github_content,
+        parse_github_content,
+    )
+
     # get current time to calculate total time taken
     start_time = time.time()
-    try: 
+    try:
         client = httpx.AsyncClient()
         logger.info(
             f"  Started     Proc.handle_github_url(): github_url - {github_url}"
@@ -97,6 +111,20 @@ async def handle_github_url(github_url: str) -> dict:
                 {"github_url": github_url},
                 500,
             )
+    except GithubError as exc:
+        proc_exception(
+            exc.source,
+            exc.message,
+            exc.params,
+            exc,
+        )
+    except ImportSourceUrlError as exc:
+        proc_exception(
+            exc.source,
+            exc.message,
+            exc.params,
+            exc,
+        )
     except Exception as exc:
         proc_exception(
             "handle_github_url",
@@ -104,7 +132,6 @@ async def handle_github_url(github_url: str) -> dict:
             {"github_url": github_url},
             exc,
         )
-
     finally:
         await client.aclose()
         logger.info(f"  Finished    Proc.handle_github_url()")
@@ -113,177 +140,3 @@ async def handle_github_url(github_url: str) -> dict:
         total_time = time.strftime("%H:%M:%S", time.gmtime(end_time - start_time))
         logger.info(f"  Total time taken: {total_time}")
         # TODO: Log this in database later
-        
-
-
-async def read_github_content(
-    url: str,
-    owner: str,
-    repo: str,
-    path: str = "",
-    first: bool = False,
-) -> list[dict] | dict:
-    try:
-        client = httpx.AsyncClient()
-
-        # Construct the API URL
-        with open("/run/secrets/github_token", "r") as file:
-            GIT_API_KEY = file.read().strip() 
-        if not GIT_API_KEY or GIT_API_KEY == "":
-            return proc_error(
-                "read_github_content",
-                "No GitHub API key found",
-                {"url": url, "api_url": "nothing"},
-                403,
-            )
-        headers = {
-            "Accept": "application/vnd.github.v3+json",
-            # Uncomment and set your token if you have one
-            "Authorization": f"token {GIT_API_KEY}",
-        }
-
-        # get the size of the whole repo
-        if first:
-            api_url = f"https://api.github.com/repos/{owner}/{repo}"
-            response = await client.get(
-                api_url, headers=headers, follow_redirects=False
-            )
-            if response.status_code == 200:
-                json_data = response.json()
-                if json_data["size"]:
-                    size = json_data["size"]
-                    logger.info(f"  Repo Size: {size} bytes")
-                    if size > 1000000:
-                        return proc_error(
-                            "read_github_content",
-                            "GitHub repo is too large",
-                            {"url": url, "api_url": api_url},
-                            500,
-                        )
-
-        # get the actual contents of the repo
-        api_url = f"https://api.github.com/repos/{owner}/{repo}/contents/{path}"
-        response = await client.get(api_url, headers=headers, follow_redirects=False)
-
-        # Check the response
-        if response.status_code == 200:
-            json_data = response.json()
-            if not json_data:
-                return proc_error(
-                    "read_github_content",
-                    "No data returned from GitHub API for URL",
-                    {"url": url, "api_url": api_url},
-                    404,
-                )
-            else:
-                # Remove unnecessary data from the response
-                # this will leave us with {name, path, size, html_url, download_url, type}
-                # html url is the url to view the file in the browser
-                # download url is the url to see just the raw file contents
-                for item in json_data:
-                    item.pop("sha", None)
-                    item.pop("url", None)
-                    item.pop("git_url", None)
-                    item.pop("_links", None)
-
-                return json_data
-        else:
-            if response.status_code == 404:
-                return proc_error(
-                    "read_github_content",
-                    "GitHub API returned 404",
-                    {"url": url, "api_url": api_url},
-                    404,
-                )
-            if response.status_code == 403:
-                error_message = f"GitHub API returned 403: {response.text}"
-                if "rate_limit" in response.text:
-                    error_message = f"GitHub API rate limit exceeded: {response.text}"
-                return proc_error(
-                    "read_github_content",
-                    error_message,
-                    {"url": url, "api_url": api_url},
-                    403,
-                )
-            else:
-                return proc_error(
-                    "read_github_content",
-                    "Error with client response",
-                    {"url": url, "status_code": response.status_code},
-                    500,
-                )
-    except httpx.RequestError as exc:
-        proc_exception(
-            "read_github_content",
-            "Error while attempting to set up request url & headers",
-            {"url": url},
-            exc,
-        )
-    except Exception as exc:
-        proc_exception(
-            "read_github_content",
-            "Error when reading GitHub content",
-            {"url": url},
-            exc,
-        ) 
-        
-
-
-async def parse_github_content(file_content, owner, repo) -> dict:
-    try:
-        # Check that the file content is a list
-        if not file_content or not isinstance(file_content, list):
-            return proc_error(
-                "parse_github_content",
-                "Invalid file content format",
-                {},
-                404,
-            )
-
-        # Process directories
-        results = {}
-        directories: list = [item for item in file_content if item["type"] == "dir"]
-        files: list = [item for item in file_content if item["type"] == "file"]
-
-        for dir in directories:
-            dir_content = await read_github_content("", owner, repo, dir["path"])
-            parsed_dir_content = await parse_github_content(dir_content, owner, repo)
-            dir_name = dir["name"]
-            results[dir_name] = parsed_dir_content
-
-        # Process files
-        top_files = []
-        for file in files:
-            top_files.append(file)
-        if top_files and len(top_files) > 0:
-            results["files"] = top_files
-
-        return results
-    except Exception as exc:
-        proc_exception(
-            "parse_github_content",
-            "Error when parsing GitHub content",
-            {"owner": owner, "repo": repo},
-            exc,
-        ) 
-        
-
-
-def raw_to_graph(raw_data: str, filename: str):
-    try:
-        logger.info(f"  Started     Proc.text_to_json()")
-        from src.parser.parser import Parser
-
-        parser = Parser(None, {"raw": raw_data, "filename": filename})
-        graph = parser.graph
-        return graph
-    except Exception as exc:
-        proc_exception(
-            "text_to_json",
-            "Error when transforming raw data to JSON",
-            {},
-            exc,
-        )
-    finally:
-        logger.info(f"  Finished    Proc.text_to_json()")
-
